Moved database handling into separate file from CameraFragment
